% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifiers.R
\name{aggregator}
\alias{aggregator}
\title{Aggregate Logical Vector Values}
\usage{
aggregator(logical_vec, type = NULL, na.rm = FALSE)
}
\arguments{
\item{logical_vec}{A logical vector to be aggregated.}

\item{type}{A character string indicating the type of aggregation.
Possible values: "p" for proportions, "n" for counts.}

\item{na.rm}{Logical. Should missing values be removed when aggregating?}
}
\value{
Depending on the \code{type} parameter, returns either the proportion
of \code{TRUE} values in the logical vector or the count of
\code{TRUE} values.
}
\description{
This internal function aggregates logical vector values based on the
specified type of aggregation requested (proportion of summation).
It is used by the package internally and not meant to be directly accessed by
users.
}
\examples{
logical_vector <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
aggregator(logical_vector, type = "p")
aggregator(logical_vector, type = "n")

}
\keyword{internal}
