% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{constructor}
\alias{constructor}
\title{Generic function for checking that batteries are all NA or not NA}
\usage{
constructor(data, fun, ...)
}
\description{
Generic function for checking that batteries are all NA or not NA
}
\keyword{internal}
