% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifiers.R
\name{is_p_or_n}
\alias{is_p_or_n}
\title{Determine if a Value Represents a Proportion (p) or Count (n)}
\usage{
is_p_or_n(num, error_on_1 = TRUE, error_on_nothing = TRUE)
}
\arguments{
\item{num}{A numeric value to be evaluated.}

\item{error_on_1}{Logical. Should an error be raised if \code{num} is equal
to 1? Defaults to \code{TRUE}. This is important in cases
where 1 is an ambiguous case (e.g., it could represent
a probability of exactly 1.0 or a count of exactly 1).}

\item{error_on_nothing}{Logical. Should an error be raised if \code{num} is
not a valid proportion (p) or count (n)? Defaults to
\code{TRUE}.}
}
\value{
A character string: "p" if \code{num} represents a proportion, "n" if
it represents a count, and \code{FALSE} if neither applies.
}
\description{
This internal function determines if a given value represents a proportion
(p) or a count (n). It is used by the package internally to handle cases
where a value can be interpreted as either a proportion or a count. This
function helps ensure consistent and safe handling of such values and is used
as a check in several other package functions.
}
\examples{
# Determine if 0.5 is a proportion (p) or a count (n)
is_p_or_n(0.5)  # Returns "p"

# Determine if 10 is a proportion (p) or a count (n)
is_p_or_n(10)  # Returns "n"

# Determine if 1 is a proportion (p) or a count (n) and allow error on 1
is_p_or_n(1, error_on_1 = TRUE)  # Throws an error

# Determine if 1 is a proportion (p) or a count (n) and allow error on nothing
is_p_or_n(1, error_on_nothing = TRUE)  # Throws an error

}
\keyword{internal}
