% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typechecks.R
\name{is_proportion}
\alias{is_proportion}
\title{Check if a Numeric Value is a Proportion}
\usage{
is_proportion(x)
}
\arguments{
\item{x}{Numeric value to check.}
}
\value{
\code{TRUE} if \code{x} is a proportion, otherwise \code{FALSE}.
}
\description{
This function checks if a numeric value is a proportion scalar, meaning it
is numeric and within the range of 0 to 1 (inclusive).
}
\examples{
is_proportion(0.5) # TRUE
is_proportion(1.2) # FALSE
is_proportion(-0.2) # FALSE
}
\seealso{
\code{\link{is_integerlike}}, \code{\link{is_count}},
\code{\link{validate_proportion}}, \code{\link{validate_count}}
}
