% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_match.R
\name{levels_match}
\alias{levels_match}
\title{Check that the levels of a group of variables match a reference set}
\usage{
levels_match(data, vars, ref_levels, strictness = "identical")
}
\description{
It's important that Tech Bullseye battery items all contain the same
response options.
}
