% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifiers.R
\name{prop}
\alias{prop}
\title{Placeholder}
\usage{
prop(logical_vec, na.rm = FALSE)
}
\arguments{
\item{logical_vec}{Another thought}

\item{na.rm}{a third thought}
}
\value{
a value
}
\description{
Here is a thought
}
