% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifiers.R
\name{at_least}
\alias{at_least}
\title{At least}
\usage{
at_least(logical_vec, p = NULL, n = NULL, na.rm = FALSE)
}
\arguments{
\item{logical_vec}{a}

\item{p}{a}

\item{n}{b}

\item{na.rm}{c}
}
\value{
d
}
\description{
xxx
}
