% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkthat-package.R
\docType{package}
\name{checkthat-package}
\alias{checkthat}
\alias{checkthat-package}
\title{checkthat: Implements Unit Testing For Data Manipulation Pipelines}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iancero/checkthat}
  \item \url{https://iancero.github.io/checkthat/}
  \item Report bugs at \url{https://github.com/iancero/checkthat/issues}
}

}
\author{
\strong{Maintainer}: Ian Cero \email{ian_cero@urmc.rochester.edu} (\href{https://orcid.org/0000-0002-2862-0450}{ORCID})

}
\keyword{internal}
