% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkthat-package.R
\docType{package}
\name{checkthat-package}
\alias{checkthat}
\alias{checkthat-package}
\title{checkthat: Intuitive Unit Testing Tools For Data Wrangling Pipelines}
\description{
The checkthat package provides a lightweight data validation and testing toolkit for R. Its guiding philosophy is that adding code-based data checks to users' existing workflow should be both quick and intuitive. The suite of functions included therefore mirror the common data checks many users already perform by hand or by eye. Additionally, the checkthat package is optimized to work within tidyverse data manipulation pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iancero/checkthat}
  \item \url{https://iancero.github.io/checkthat/}
  \item Report bugs at \url{https://github.com/iancero/checkthat/issues}
}

}
\author{
\strong{Maintainer}: Ian Cero \email{ian_cero@urmc.rochester.edu} (\href{https://orcid.org/0000-0002-2862-0450}{ORCID}) [copyright holder]

}
\keyword{internal}
