% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typechecks.R
\name{validate_proportion}
\alias{validate_proportion}
\title{Validate a Proportion Value}
\usage{
validate_proportion(x)
}
\arguments{
\item{x}{Numeric value to validate as a proportion.}
}
\value{
\code{TRUE} if \code{x} is a valid proportion, otherwise it throws
an error.
}
\description{
This function validates whether a numeric value is a valid proportion scalar
(ranging from 0 to 1, inclusive).
}
\examples{
validate_proportion(0.5)   # TRUE
try(validate_proportion(1.2))   # Error
}
\seealso{
\code{\link{is_proportion}}, \code{\link{is_count}},
\code{\link{validate_count}}, \code{\link{is_integerlike}}
}
